% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correctImpurities.MaxQuant}
\alias{correctImpurities.MaxQuant}
\title{Performs purity correction on MaxQuant data at specified levels
The number of intensity columns must equal the number of reporters in the impurity file
Output is a list of data frames for each level of analysis}
\usage{
correctImpurities.MaxQuant(
  impurities,
  txt.path,
  levels = c("msms", "evidence", "proteinGroups"),
  method = c("NNLS", "OLS"),
  noise.replacement.method = c("pre", "post", "none"),
  remain.missing = T,
  remove.missing.rows = T,
  use.razor = T,
  guess_max = 20000
)
}
\arguments{
\item{impurities}{data frame of impurities for a specific lot}

\item{txt.path}{path to the MaxQuant txt output folder. If using noiseband imputation, it should contain an intensity.txt and noise.txt file.}

\item{levels}{list of which levels to perform purity correction and noiseband imputation. Options are "msms", "evidence", and "proteinGroups"}

\item{method}{purity correction method. Options = "NNLS" and "OLS"}

\item{noise.replacement.method}{should the noiseband replace missing values before or after correction. Options = "pre", "post", "none"}

\item{remain.missing}{should missing values be set to 0 (missing) after correction or not. Only used if not doing noiseband imputation. True or False}

\item{remove.missing.rows}{should rows containing all missing rows be removed or not. True or False}

\item{use.razor}{should razor peptides be used for protein quantification. Only used for proteinGroups and if msms is supplied.}

\item{guess_max}{guess_max used for reading the csv files. It's the number of rows to read to guess column data types.}
}
\description{
Performs purity correction on MaxQuant data at specified levels
The number of intensity columns must equal the number of reporters in the impurity file
Output is a list of data frames for each level of analysis
}
