% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correctImpurities.proteinGroups}
\alias{correctImpurities.proteinGroups}
\title{Performs purity correction on MaxQuant proteinGroups.txt data}
\usage{
correctImpurities.proteinGroups(
  summary,
  proteinGroups,
  impurities,
  msms,
  intensities,
  noise,
  use.razor = T,
  method = c("NNLS", "OLS"),
  noise.replacement.method = c("pre", "post", "none"),
  remain.missing = T,
  remove.missing.rows = T
)
}
\arguments{
\item{summary}{data frame of summary.txt}

\item{proteinGroups}{data frame of evidence.txt}

\item{impurities}{data frame of impurities for a specific lot}

\item{msms}{optional data frame of msms.txt to perform correction at spectrum level. Required for noiseband imputation.}

\item{intensities}{data frame of TMT intensity values, raw file, and MS2 scan number}

\item{noise}{optional data frame of TMT noiseband values, raw file, and MS2 scan number}

\item{use.razor}{should razor peptides be used for protein quantification. Only used if msms is supplied.}

\item{method}{purity correction method. Options = "NNLS" and "OLS"}

\item{noise.replacement.method}{should the noiseband replace missing values before or after correction. Options = "pre", "post", and "none"}

\item{remain.missing}{should missing values be set to 0 (missing) after correction or not. Only used if not doing noiseband imputation. True or False}

\item{remove.missing.rows}{should rows containing all missing rows be removed or not. True or False}
}
\description{
Performs purity correction on MaxQuant proteinGroups.txt data
}
\examples{
\dontrun{

}


}
