% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correctImpurities}
\alias{correctImpurities}
\title{Performs purity correction on generic TMT data.}
\usage{
correctImpurities(
  data,
  impurities,
  noise,
  method = c("NNLS", "OLS"),
  noise.replacement.method = c("pre", "post", "none"),
  remain.missing = T
)
}
\arguments{
\item{data}{data frame of uncorrected TMT intensities.}

\item{impurities}{data frame of impurities for a specific lot}

\item{noise}{optional data frame of TMT noiseband values. Must line up row by row with data}

\item{method}{purity correction method. Options = "NNLS" and "OLS"}

\item{noise.replacement.method}{should the noiseband replace missing values before or after correction. Options = "pre", "post", "none"}

\item{remain.missing}{should missing values be set to 0 (missing) after correction or not. Only used if not doing noiseband imputation. True or False}
}
\description{
Performs purity correction on generic TMT data.
}
\examples{
\dontrun{

}


}
