% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correctImpurities.RTS}
\alias{correctImpurities.RTS}
\title{Performs purity correction on TMT RTS data.
The number of intensity columns must equal the number of reporters in the impurity file
Output is a data frame of the same dimensions}
\usage{
correctImpurities.RTS(
  msms,
  impurities,
  noise,
  method = c("NNLS", "OLS"),
  noise.replacement.method = c("pre", "post"),
  remain.missing = T,
  remove.missing.rows = T,
  use.razor = T
)
}
\arguments{
\item{msms}{data frame of uncorrected TMT intensities, raw file, and MS2 scan number}

\item{impurities}{data frame of impurities for a specific lot}

\item{noise}{optional data frame of TMT noiseband values, raw file, and MS2 scan number}

\item{method}{purity correction method. Options = "NNLS" and "OLS"}

\item{noise.replacement.method}{should the noiseband replace missing values before or after correction. Options = "pre" and "post"}

\item{remain.missing}{should missing values be set to 0 (missing) after correction or not. Only used if not doing noiseband imputation. True or False}

\item{remove.missing.rows}{should rows containing all missing rows be removed or not. True or False}
}
\description{
Performs purity correction on TMT RTS data.
The number of intensity columns must equal the number of reporters in the impurity file
Output is a data frame of the same dimensions
}
\examples{
\dontrun{
}


}
