% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_impurities}
\alias{default_impurities}
\title{Impurities for TMTpro lot VJ309267}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16 rows and 18 columns.
}
\usage{
default_impurities
}
\description{
Impurities for TMTpro lot VJ309267
}
\keyword{datasets}
